% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_paths.R
\name{air_exec}
\alias{air_exec}
\title{Path to 'air' executable}
\usage{
air_exec()
}
\value{
executable full path in a character vector of length 1.
}
\description{
return the full path of 'air' executable
(the R formatter and language server) if found.
}
\examples{
if(exec_available("air"))
  message(air_exec())
}
\seealso{
Other executable full path:
\code{\link[=chrome_exec]{chrome_exec()}},
\code{\link[=excel_exec]{excel_exec()}},
\code{\link[=firefox_exec]{firefox_exec()}},
\code{\link[=libreoffice_exec]{libreoffice_exec()}},
\code{\link[=node_exec]{node_exec()}},
\code{\link[=npm_exec]{npm_exec()}},
\code{\link[=pandoc_exec]{pandoc_exec()}},
\code{\link[=pip_exec]{pip_exec()}},
\code{\link[=powerpoint_exec]{powerpoint_exec()}},
\code{\link[=python_exec]{python_exec()}},
\code{\link[=word_exec]{word_exec()}}
}
\concept{executable full path}
