% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_paths.R
\name{chrome_exec}
\alias{chrome_exec}
\title{Path to 'Google Chrome' executable}
\usage{
chrome_exec()
}
\value{
chrome executable full path in a character vector of length 1.
}
\description{
return the full path of 'Google Chrome' executable
if found.
}
\examples{
if(exec_available("chrome"))
  message(chrome_exec())
}
\seealso{
Other executable full path: 
\code{\link{excel_exec}()},
\code{\link{firefox_exec}()},
\code{\link{libreoffice_exec}()},
\code{\link{node_exec}()},
\code{\link{npm_exec}()},
\code{\link{pip_exec}()},
\code{\link{powerpoint_exec}()},
\code{\link{python_exec}()},
\code{\link{word_exec}()}
}
\concept{executable full path}
