% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_paths.R
\name{firefox_exec}
\alias{firefox_exec}
\title{Path to 'Mozilla Firefox' executable}
\usage{
firefox_exec()
}
\value{
Firefox executable full path in a character vector of length 1.
}
\description{
return the full path of 'Mozilla Firefox' executable
if found.
}
\examples{
if(exec_available("firefox"))
  message(firefox_exec())
}
\seealso{
Other executable full path: 
\code{\link{chrome_exec}()},
\code{\link{excel_exec}()},
\code{\link{libreoffice_exec}()},
\code{\link{node_exec}()},
\code{\link{npm_exec}()},
\code{\link{pip_exec}()},
\code{\link{powerpoint_exec}()},
\code{\link{python_exec}()},
\code{\link{word_exec}()}
}
\concept{executable full path}
