% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_os.R
\name{is_osx}
\alias{is_osx}
\title{Test if the system is "macOS"}
\usage{
is_osx()
}
\value{
TRUE or FALSE
}
\description{
Test if the operating system is "macOS"
}
\examples{
is_osx()
}
\seealso{
Other operating system testing: 
\code{\link{is_unix}()},
\code{\link{is_windows}()}
}
\concept{operating system testing}
