% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_os.R
\name{is_unix}
\alias{is_unix}
\title{Test if the system is "unix"}
\usage{
is_unix()
}
\value{
TRUE or FALSE
}
\description{
Test if the operating system is "unix"
}
\examples{
is_unix()
}
\seealso{
Other operating system testing: 
\code{\link{is_osx}()},
\code{\link{is_windows}()}
}
\concept{operating system testing}
