% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_os.R
\name{is_windows}
\alias{is_windows}
\title{Test if the system is "Windows"}
\usage{
is_windows()
}
\value{
TRUE or FALSE
}
\description{
Test if the operating system is "Windows"
}
\examples{
is_windows()
}
\seealso{
Other operating system testing: 
\code{\link{is_osx}()},
\code{\link{is_unix}()}
}
\concept{operating system testing}
