% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_paths.R
\name{node_exec}
\alias{node_exec}
\title{Path to 'node.js' executable}
\usage{
node_exec()
}
\value{
executable full path in a character vector of length 1.
}
\description{
return the full path of 'node.js' executable
if found.
}
\examples{
if(exec_available("node"))
  message(node_exec())
}
\seealso{
Other executable full path: 
\code{\link{chrome_exec}()},
\code{\link{excel_exec}()},
\code{\link{firefox_exec}()},
\code{\link{libreoffice_exec}()},
\code{\link{npm_exec}()},
\code{\link{pip_exec}()},
\code{\link{powerpoint_exec}()},
\code{\link{python_exec}()},
\code{\link{word_exec}()}
}
\concept{executable full path}
