% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_paths.R
\name{word_exec}
\alias{word_exec}
\title{Path to 'Microsoft Word' executable}
\usage{
word_exec()
}
\value{
executable full path in a character vector of length 1.
}
\description{
return the full path of 'Microsoft Word' executable
if found.
}
\examples{
if(exec_available("word"))
  message(word_exec())
}
\seealso{
Other executable full path: 
\code{\link{chrome_exec}()},
\code{\link{excel_exec}()},
\code{\link{firefox_exec}()},
\code{\link{libreoffice_exec}()},
\code{\link{node_exec}()},
\code{\link{npm_exec}()},
\code{\link{pip_exec}()},
\code{\link{powerpoint_exec}()},
\code{\link{python_exec}()}
}
\concept{executable full path}
